Working on reshaping data
In my first use of this Keras model the data was shaped (2733, ) but with how I don't crop the data and fuse it back together its now (2733, 224, 224) which doesn't work.  So either I need to figure out how to reshape the data or rebuild the CNN to work with what I have.